requirements.txt
gspread
google-auth
beautifulsoup4
requests